Changed YouCompleteMe for SuperTab again.
Although the fuzzy completion of YCM is absolutely awesome, it's getting really slow for me to edit remote java files. This is because YCM tries to auto complete while I type and this triggers the Eclim completion.

A branch for who likes the YCM was created and I'll try to keep it updated together with 'master'.

Change-Id: I77e99ab30dbcaee5c83c771b859f7a48c782a256