Fix Greper mappings.
[Problem]
When pressing ,sa in normal mode or visual mode, the Greper command failed. On
Graper documentation the author state that using nnoremap to map its utilities
does not work and I was using exactly that.

[Solution]
Change the nnoremap to nmap on Greper mappings.
Also, I changed the Greper call in visual mode to use the /patern/ search
instead of "patern".