Fix Rooter paterns.
[Problem]
I was using a to broad condition to define the root of a project and this was
causing Rooter to believe that a child directory should be considered a root one
(more specifically Android projects).

[Solution]
Only consider root directory if we find a build script file (Makefile,
build.gradle, etc.) in it. Also, we check for VCS patters like .git, .hg, etc.