Massive change on my configureation, read on for details.
After a long consideration I decided to use vimPlug as my plugin manager and
among other things, this change is to reflect that decision.

Unfortunately during this process I had to to do tons of tweaks on my config and
I was not diligent enough to commit those change as I was doing them. So, since
the effort to separate all those commits is too big and the risk of committing
something that breaks the config without the next commit is also too big, I
decided to do a big commit to my vimrc and move on.

I'll try to make the next tweaks more granular to everyone that cares about my
configuration.