Fix the <F3> problem on my terminal.
[Problem]
For reasons I couldn't find out, when I was inside an Unite buffer (most
notorious the buffer explorer one) and I press the <F3> key, a new line
containing only the character 'R' was inserted on my document.

It turns out that the <F3> key is internally mapped to a combination of keys for
vt100 terminals: <ESC>OR

This caused, first the Unite buffer to close with the <ESC> key, then a new line
was inserted above my current one with the key 'O' and after that the letter 'R'
was inserted.

[Solution]
This is not a solution to the problem itself but it also prevent it from
happening. I simple map, only inside an Unite buffer, the sequence of keys that
<F3> (<ESC>OR) sends to the same "toggle" command that <F3> originally was
supposed to send.